refactor: tests. added asserts, made them stable

+++ src/main/java/io/tatum/model/response/tron/Trc10.java
@@ -6,6 +6,8 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
@@ -13,5 +15,5 @@ import lombok.NoArgsConstructor;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Trc10 {
     private String key;
-    private  long value;
+    private BigDecimal value;
 }
